Ajustes en invitation

+++ src/apps/coopolis/views/ProjectFormView.py
@@ -75,10 +75,11 @@ class ProjectFormView(SuccessMessageMixin, generic.UpdateView):
 
 @login_required
 def project_partner_manage(request):
+    project_pk = None
     if request.method == "POST":
-        project_pk = request.POST.get('add_partner')
-        project = Project.objects.get(pk=project_pk)
         if "add_partner" in request.POST:
+            project_pk = request.POST.get("add_partner")
+            project = Project.objects.get(pk=project_pk)
             validator = EmailValidator()
             try:
                 validator(request.POST.get("email"))
@@ -93,9 +94,14 @@ def project_partner_manage(request):
             except User.DoesNotExist:
                 messages.error(
                     request,
-                    "L'usuari no existeix.", )
+                    "L'usuari no existeix.",
+                )
                 return redirect("edit_project", pk=project_pk)
-            user_is_invited = Invitation.objects.filter(user=user, project=project).values("user__id").first()
+            user_is_invited = (
+                Invitation.objects.filter(user=user, project=project)
+                .values("user__id")
+                .first()
+            )
             if user_is_invited:
                 messages.error(
                     request,
@@ -117,7 +123,7 @@ def project_partner_manage(request):
                 "persona_email": user.email,
                 "project": project.name,
                 "absolute_url": settings.ABSOLUTE_URL,
-                "invitation_url": f"{settings.ABSOLUTE_URL}/project/invitation/{str(invitation.uuid)}"
+                "invitation_url": f"{settings.ABSOLUTE_URL}/project/invitation/{str(invitation.uuid)}",
             }
             mail.send(now=True)
             messages.success(
@@ -126,6 +132,15 @@ def project_partner_manage(request):
                 " Se li ha enviat un correu per a la seva acceptació.",
             )
         if "delete_partner" in request.POST:
+            project_pk = request.POST.get("delete_partner")
+            try:
+                project = Project.objects.get(pk=project_pk)
+            except Project.DoesNotExist:
+                messages.error(
+                    request,
+                    "Aquest projecte no existeix.",
+                )
+                return redirect("home")
             try:
                 user = User.objects.get(id=request.POST.get("partner_id"))
             except User.DoesNotExist:
@@ -133,21 +148,13 @@ def project_partner_manage(request):
                     request,
                     "Aquest usuari no existeix.",
                 )
-                return redirect("edit_project", pk=project_pk)
+                return redirect("edit_project", pk=project.pk)
             if user == request.user:
                 messages.error(
                     request,
                     "No et pots eliminar a tu mateix/a del projecte.",
                 )
-                return redirect("edit_project", pk=project_pk)
-            try:
-                project = Project.objects.get(pk=request.POST.get("delete_partner"))
-            except Project.DoesNotExist:
-                messages.error(
-                    request,
-                    "Aquest projecte no existeix.",
-                )
-                return redirect("edit_project", pk=project_pk)
+                return redirect("edit_project", pk=project.pk)
             project.partners.remove(user)
 
             mail = MyMailTemplate("EMAIL_PARTNER_ELIMINATION")
@@ -161,32 +168,34 @@ def project_partner_manage(request):
             mail.send(now=True)
             messages.success(
                 request,
-                f"{user.full_name} ha estat eliminada amb èxit d'aquest projecte.", )
-
-    if "delete_invitation" in request.POST:
-        try:
-            user = User.objects.get(id=request.POST.get("invited_id"))
-        except User.DoesNotExist:
-            messages.error(
-                request,
-                "Aquest usuari no existeix.",
+                f"{user.full_name} ha estat eliminada amb èxit d'aquest projecte.",
             )
-            return redirect("edit_project", pk=project_pk)
-        try:
-            invitation = Invitation.objects.get(user=user, project=request.POST.get("delete_invitation"))
-        except Invitation.DoesNotExist:
-            messages.error(
+        if "delete_invitation" in request.POST:
+            project_pk = request.POST.get("delete_invitation")
+            try:
+                user = User.objects.get(id=request.POST.get("invited_id"))
+            except User.DoesNotExist:
+                messages.error(
+                    request,
+                    "Aquest usuari no existeix.",
+                )
+                return redirect("edit_project", pk=project_pk)
+            try:
+                invitation = Invitation.objects.get(
+                    user=user, project=request.POST.get("delete_invitation")
+                )
+            except Invitation.DoesNotExist:
+                messages.error(
+                    request,
+                    "Aquesta invitació no existeix.",
+                )
+                return redirect("edit_project", pk=project_pk)
+            invitation.delete()
+            messages.success(
                 request,
-                "Aquesta invitació no existeix.",
+                f"{user.full_name} ha estat eliminada amb èxit d'aquest projecte.",
             )
-            return redirect("edit_project", pk=project_pk)
-        invitation.delete()
-        messages.success(
-            request,
-            f"{user.full_name} ha estat eliminada amb èxit d'aquest projecte.",
-        )
     return redirect("edit_project", pk=project_pk)
-
 
 
 @login_required
@@ -199,20 +208,18 @@ def invitation_partner(request, uuid):
                 request,
                 "Aquest enllaç d'invitació no existeix.",
             )
-            return redirect("edit_project")
+            return redirect("home")
         if request.user.id != invitation.user.id:
             messages.error(
                 request,
                 "Estàs accedint a un enllaç d'invitació a un projecte que correspon a un altre usuari. "
                 "Assegura't d'obrir l'enllaç des del compte a la qual t'han enviat la invitació.",
             )
-            return redirect("edit_project")
-        context = {
-            "project": invitation.project,
-            "user": invitation.user
-        }
+            return redirect("home")
+        context = {"project": invitation.project, "user": invitation.user}
         return render(request, "project_invitation.html", context=context)
     if request.method == "POST":
+        project_pk = request.POST.get("project")
         try:
             invitation = Invitation.objects.get(uuid=uuid)
         except Invitation.DoesNotExist:
@@ -220,7 +227,7 @@ def invitation_partner(request, uuid):
                 request,
                 "Aquest enllaç d'invitació no existeix.",
             )
-            return redirect("edit_project")
+            return redirect("home")
         try:
             user = User.objects.get(id=invitation.user.id)
         except User.DoesNotExist:
@@ -228,19 +235,31 @@ def invitation_partner(request, uuid):
                 request,
                 "Aquest usuari no existeix.",
             )
-            return redirect("edit_project")
+            return redirect("home")
         if "accept" in request.POST:
             try:
-                project = Project.objects.get(pk=request.POST.get("project"))
+                project = Project.objects.get(pk=project_pk)
             except Project.DoesNotExist:
                 messages.error(
                     request,
                     "Aquest projecte no existeix.",
                 )
-                return redirect("edit_project")
+                return redirect("home")
             project.partners.add(user)
-        invitation.delete()
-        return redirect("edit_project")
+            invitation.delete()
+            messages.error(
+                request,
+                "Has acceptat formar part del projecte.",
+            )
+            return redirect("edit_project", pk=project_pk)
+        if "deny" in request.POST:
+            invitation.delete()
+            messages.error(
+                request,
+                "Has rebutjat formar part del projecte.",
+            )
+            return redirect("home")
+    return redirect("home")
 
 
 class ProjectCreateFormView(SuccessMessageMixin, generic.CreateView):
